+++ pennylane_lightning/lightning_tensor/lightning_tensor.py
@@ -22,7 +22,7 @@ from warnings import warn
 
 import numpy as np
 import pennylane as qml
-from pennylane.devices import DefaultExecutionConfig, Device, ExecutionConfig
+from pennylane.devices import Device, ExecutionConfig
 from pennylane.devices.modifiers import simulator_tracking, single_tape_support
 from pennylane.devices.preprocess import (
     decompose,
@@ -402,13 +402,13 @@ class LightningTensor(Device):
 
     dtype = c_dtype
 
-    def _setup_execution_config(
-        self, config: Optional[ExecutionConfig] = DefaultExecutionConfig
-    ) -> ExecutionConfig:
+    def _setup_execution_config(self, config: Optional[ExecutionConfig] = None) -> ExecutionConfig:
         """
         Update the execution config with choices for how the device should be used and the device options.
         """
         # TODO: add options for gradients next quarter
+        if config is None:
+            config = ExecutionConfig()
 
         updated_values = {}
 
@@ -433,7 +433,7 @@ class LightningTensor(Device):
 
     def preprocess(
         self,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ):
         """This function defines the device transform program to be applied and an updated device configuration.
 
@@ -451,6 +451,8 @@ class LightningTensor(Device):
         * Does not support derivatives.
         * Does not support vector-Jacobian products.
         """
+        if execution_config is None:
+            execution_config = ExecutionConfig()
 
         config = self._setup_execution_config(execution_config)
 
@@ -472,7 +474,7 @@ class LightningTensor(Device):
     def execute(
         self,
         circuits: QuantumTape_or_Batch,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ) -> Result_or_ResultBatch:
         """Execute a circuit or a batch of circuits and turn it into results.
 
@@ -521,7 +523,7 @@ class LightningTensor(Device):
     def compute_derivatives(
         self,
         circuits: QuantumTape_or_Batch,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ):
         """Calculate the Jacobian of either a single or a batch of circuits on the device.
 
@@ -539,7 +541,7 @@ class LightningTensor(Device):
     def execute_and_compute_derivatives(
         self,
         circuits: QuantumTape_or_Batch,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ):
         """Compute the results and Jacobians of circuits at the same time.
 
@@ -575,7 +577,7 @@ class LightningTensor(Device):
         self,
         circuits: QuantumTape_or_Batch,
         cotangents: Tuple[Number],
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ):
         r"""The vector-Jacobian product used in reverse-mode differentiation.
 
@@ -597,7 +599,7 @@ class LightningTensor(Device):
         self,
         circuits: QuantumTape_or_Batch,
         cotangents: Tuple[Number],
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
+        execution_config: Optional[ExecutionConfig] = None,
     ):
         """Calculate both the results and the vector-Jacobian product used in reverse-mode differentiation.